# api/openapi.yaml
openapi: 3.0.0
info:
  title: API de Guardado
  description: API para la gestión de guardado en una plataforma de videojuegos
  version: 1.0.0
servers:
- url: https://indie-world-api-guardado.vercel.app/Devgram/ApiGuardado/1.0.0
  description: Vercel Server Production
paths:
  /datos/guardar:
    post:
      tags:
      - Datos
      summary: Guardar archivo
      description: Guarda un archivo especificado en la nube o localmente.
      operationId: guardarArchivo
      parameters:
      - name: ubicacion
        in: query
        description: Ubicación del archivo (local o nube)
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - local
          - nube
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchivoGuardado'
        required: true
      responses:
        "200":
          description: Archivo guardado correctamente
        "400":
          description: Error en la solicitud
      x-swagger-router-controller: Datos
  /datos/cargar:
    get:
      tags:
      - Datos
      summary: Cargar archivo
      description: Carga un archivo desde la nube o localmente.
      operationId: cargarArchivo
      parameters:
      - name: id_archivo_guardado
        in: query
        description: ID del archivo a cargar
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      - name: ubicacion
        in: query
        description: Ubicación del archivo (local o nube)
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - local
          - nube
      responses:
        "200":
          description: Archivo cargado correctamente
        "404":
          description: Archivo no encontrado
        "400":
          description: Error en la solicitud
      x-swagger-router-controller: Datos
  /datos/eliminar:
    delete:
      tags:
      - Datos
      summary: Eliminar archivo
      description: Elimina un archivo de la nube o localmente.
      operationId: eliminarArchivo
      parameters:
      - name: id
        in: query
        description: ID del archivo a eliminar
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      - name: ubicacion
        in: query
        description: Ubicación del archivo (local o nube)
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - local
          - nube
      responses:
        "204":
          description: Archivo eliminado correctamente
        "404":
          description: Archivo no encontrado
        "400":
          description: Error en la solicitud
      x-swagger-router-controller: Datos
  /metrics:
    get:
      tags:
      - Datos
      summary: Obtiene metricas de la api
      description: Obtiene metricas de la api
      operationId: getMetrics
      responses:
        "201":
          description: item created
        "400":
          description: "invalid input, object invalid"
      x-swagger-router-controller: Datos
components:
  schemas:
    ArchivoGuardado:
      required:
      - datos
      - id_archivo_guardado
      - id_juego
      - id_usuario
      type: object
      properties:
        id_archivo_guardado:
          type: integer
          format: int64
        id_juego:
          type: integer
          format: int64
        id_usuario:
          type: integer
          format: int64
        datos:
          type: array
          items:
            $ref: '#/components/schemas/ArchivoGuardado_datos'
    ArchivoGuardado_datos:
      type: object
      properties:
        clave:
          type: string
        valor:
          type: string